Validations added

+++ tests/books.test.js
@@ -2,60 +2,75 @@ const supertest = require("supertest");
 const mongoose = require("mongoose");
 const { app, server } = require("../index");
 const Book = require("../models/book.model");
+const User = require("../models/user.model");
 
 const api = supertest(app);
 
+const user = {
+    name: "Jhon",
+    secondName: "Doe",
+    email: "jhon.doe@mail.com",
+    password: "JhonDo3!",
+};
+
+const book = {
+    title: "Harry Potter and the the filosofal stone",
+    author: "J. K. Rowling",
+    isbn: 9788478884452,
+};
+
 describe("Testing books endpoints", () => {
+    let accesssToken = "";
+
     beforeEach(async () => {
         await Book.deleteMany({});
+        await User.deleteMany({});
 
+        const tuser = new User(user);
         const tbook = new Book({
-            author: "William Gibson",
-            title: "Neuromancer",
-            isbn: 9780441000685,
-            image: "https://pictures.abebooks.com/isbn/9780441000685-es.jpg",
+            title: "The Grapes of Wrath",
+            isbn: "9780804137785",
+            description: "A book about the Grapes of Wrath",
+            author: "John Steinbeck",
         });
 
-        tbook.save();
+        await tbook.save();
+        await tuser.save();
+
+        const tokens = await api.post("/auth/signin").send({
+            email: user.email,
+            password: user.password,
+        });
+
+        accesssToken = tokens.body.accessToken;
     });
 
-    test.skip("Respons in json", async () => {
-        await api
-            .get("/books")
-            .expect(201)
-            .expect("Content-Type", /application\/json/);
+    test("Create a book", async () => {
+        await api.post("/books").send(book).set("Authorization", accesssToken).expect(201);
     });
 
-    test.skip("Create a book", async () => {
+    test("Get book detail", async () => {
+        await api.get(`/books/9780804137785`).expect(201);
+    });
+
+    test("Update a book", async () => {
         await api
-            .post("/books")
-            .send({
-                title: "Harry Potter and the the filosofal stone",
-                author: "J. K. Rowling",
-                isbn: 9788478884452,
-                image: "https://images-na.ssl-images-amazon.com/images/I/51ifu1aebKL._SY264_BO1,204,203,200_QL40_ML2_.jpg",
-            })
+            .put("/books/9780804137785")
+            .send({ title: "The Grapes of Wrath (Paperback)" })
+            .set("Authorization", accesssToken)
             .expect(201);
+
+        const res = await api.get("/books/9780804137785").expect(201);
+        expect(res.body.title).toBe("The Grapes of Wrath (Paperback)");
     });
 
-    test.skip("Get book detail", async () => {
-        const res = await api.get("/books/9780441000685").expect(201);
-        expect(res.body.isbn).toBe(9780441000685);
-    });
-
-    test.skip("Update a book", async () => {
-        await api.put("/books/9780441000685").send({ title: "Neuromancer (Paperback)" }).expect(201);
-        const res = await api.get("/books/9780441000685");
-        expect(res.body.title).toBe("Neuromancer (Paperback)");
-    });
-
-    test.skip("Listing two books", async () => {
+    test("Listing two books", async () => {
         const res = await api.get("/books").expect(201);
         expect(res.body).toBeDefined();
     });
 
-    test.skip("Remove a book", async () => {
-        await api.delete("/books/9780441000685").expect(201);
+    test("Remove a book", async () => {
+        await api.delete("/books/9780441000685").set({ Authorization: accesssToken }).expect(201);
     });
 });
